+++ CHANGELOG.md
@@ -1,2 +1,4 @@
-- Ported to 26.2
-- Updated to latest Inventory Particles & MossyLib
+Physics configs generation has been significantly optimized!
+- Previously, ~1,500 items took a minute; now it takes 1–2 seconds
+- Regenerating the old cache is not required, though you can do so if you wish
+- Updated to latest Inventory Particles version